fix: correctly parse log opts as array
The inffered type by Deserialize is Option<T> which is wrong and thus
the deserialize call fails as it is called on the wrong type. If we have
an array we need to directly use the HashSet type.

Switch one of the log statements in the tests to use two flags to cover
this path.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>